+++ apps/publikator/app/components/author-form/index.tsx
@@ -18,7 +18,7 @@ import { ArrowLeft, Save } from 'lucide-react'
 import Link from 'next/link'
 import GeneralWarningErrorCallout from './warning-error-callout'
 import { ArticleContributor } from '../../../graphql/republik-api/__generated__/gql/graphql'
-import { upsertAuthor } from '../../authors/actions'
+import { upsertAuthor } from '../../authors/(actions)/upsertAuthor'
 import DeleteAuthorButton from './delete-author-button'
 import ProfileImageUpload from './profile-image-upload'
 import ProlitterisFields from './prolitteris-fields'
